Move calculateNextDecision to StateManager
